+++ tests/test_inotify_buffer.py
@@ -17,7 +17,6 @@
 from __future__ import unicode_literals
 import os
 import random
-import time
 import pytest
 from tests import tmpdir, p  # pytest magic
 from .shell import mkdir, touch, mv
@@ -37,23 +36,13 @@ def wait_for_move_event(read_event):
             return event
 
 
-def make_inotify_buffer(path, recursive=False, delay=0.5):
-    return InotifyBuffer(path.encode(), recursive=recursive, delay=delay)
-
-
-def start_inotify_buffer(path, recursive=False):
-    inotify = make_inotify_buffer(path, recursive=recursive)
-    inotify.start()
-    return inotify
-
-
 @pytest.mark.timeout(5)
 def test_move_from(p):
     mkdir(p('dir1'))
     mkdir(p('dir2'))
     touch(p('dir1', 'a'))
 
-    inotify = start_inotify_buffer(p('dir1'))
+    inotify = InotifyBuffer(p('dir1').encode())
     mv(p('dir1', 'a'), p('dir2', 'b'))
     event = wait_for_move_event(inotify.read_event)
     assert event.is_moved_from
@@ -67,7 +56,7 @@ def test_move_to(p):
     mkdir(p('dir2'))
     touch(p('dir1', 'a'))
 
-    inotify = start_inotify_buffer(p('dir2'))
+    inotify = InotifyBuffer(p('dir2').encode())
     mv(p('dir1', 'a'), p('dir2', 'b'))
     event = wait_for_move_event(inotify.read_event)
     assert event.is_moved_to
@@ -81,13 +70,13 @@ def test_move_internal(p):
     mkdir(p('dir2'))
     touch(p('dir1', 'a'))
 
-    inotify = start_inotify_buffer(p(''), recursive=True)
+    inotify = InotifyBuffer(p('').encode(), recursive=True)
     mv(p('dir1', 'a'), p('dir2', 'b'))
     frm, to = wait_for_move_event(inotify.read_event)
     assert frm.src_path == p('dir1', 'a').encode()
     assert to.src_path == p('dir2', 'b').encode()
     inotify.close()
-
+    
 
 @pytest.mark.timeout(10)
 def test_move_internal_batch(p):
@@ -98,7 +87,7 @@ def test_move_internal_batch(p):
     for f in files:
         touch(p('dir1', f))
 
-    inotify = start_inotify_buffer(p(''), recursive=True)
+    inotify = InotifyBuffer(p('').encode(), recursive=True)
 
     random.shuffle(files)
     for f in files:
@@ -115,33 +104,8 @@ def test_move_internal_batch(p):
     inotify.close()
 
 
-@pytest.mark.timeout(5)
-def test_ignore_before_start(p):
-    """
-    It will ignore events created before start.
-    """
-    inotify = make_inotify_buffer(p(''), delay=0)
-
-    touch(p('a'))
-
-    assert 0 == len(inotify._queue)
-
-
-def test_close_clean(tmpdir):
-    """
-    On InotifyBuffer.close() InotifyBuffer.read_event() is un-blocked so that
-    Inotify thread waiting for it can be closed.
-
-    This is also a test for Inotify.queue_events handling of STOP_EVENT and
-    InotifyBuffer.close() is test as side effect of Inotify.stop()
-    """
-    watch = ObservedWatch(path=tmpdir, recursive=False)
-    emitter = InotifyEmitter([], watch)
-    inotify_buffer = emitter._inotify
-    # Remove delay to speed up test.
-    inotify_buffer.delay = 0
-    emitter.start()
-
-    emitter.stop()
-    emitter.join(1)
-    assert not emitter.isAlive()
+def test_close_should_terminate_thread(p):
+    inotify = InotifyBuffer(p('').encode(), recursive=True)
+    assert inotify.is_alive()
+    inotify.close()
+    assert not inotify.is_alive()